+++ src/utils.ts
@@ -190,6 +190,7 @@ export function parseQueryString(searchParams: URLSearchParams): Partial<QuerySt
     videoH264Params: parseStringParameter(searchParams, 'videoH264Params'),
     videoH265Params: parseStringParameter(searchParams, 'videoH265Params'),
     videoAV1Params: parseStringParameter(searchParams, 'videoAV1Params'),
+    forceStereoOutput: parseBooleanParameter(searchParams, 'forceStereoOutput'),
     audioInput: parseStringParameter(searchParams, 'audioInput'),
     videoInput: parseStringParameter(searchParams, 'videoInput'),
     audioOutput: parseStringParameter(searchParams, 'audioOutput'),
@@ -765,6 +766,11 @@ export function createConnectOptions(
       connectionOptions.audioStreamingLanguageCode =
         connectionOptionsState.audioStreamingLanguageCode
     }
+  }
+  // forceStereoOutput
+  // role が sendrecv か recvonly の場合は forceStereoOutput の設定を反映する
+  if (connectionOptionsState.role !== 'sendonly' && connectionOptionsState.forceStereoOutput) {
+    connectionOptions.forceStereoOutput = connectionOptionsState.forceStereoOutput
   }
   // spotlight
   const parsedSpotlight = parseBooleanString(connectionOptionsState.spotlight)